add exception classes

+++ src/main/java/by/it/academy/shopS/dto/ProductResponse.java
@@ -3,13 +3,16 @@ package by.it.academy.shopS.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.util.UUID;
+
 @Data
 @Builder
 public class ProductResponse {
-    private int id;
+    private Long id;
     private String model;
     private String specifications;
-    private int guarantee;
-    private int price;
-    private int quantity;
+    private Integer guarantee;
+    private BigDecimal price;
+    private Integer quantity;
 }
